Fix docs+enums JC-538
