+++ libs/core/mappers/anime-query-params.mapper.ts
@@ -1,7 +1,7 @@
 import { AnimeQueryParamsDto, AnimeSortFieldDto } from '../dtos/anime-query-params-dto';
+import { AnimeTypeDto } from '../dtos/anime.dto';
+import { AnimeType } from '../models/anime';
 import { AnimeQueryParams, AnimeSortField } from '../models/anime-query-params';
-
-import { ANIME_TYPE_TO_DTO } from './anime.mapper.';
 
 /** Anime sort field transformation object in dto. */
 const ANIME_SORT_FIELD_TO_DTO = {
@@ -10,6 +10,17 @@ const ANIME_SORT_FIELD_TO_DTO = {
 	[AnimeSortField.AiredStart]: AnimeSortFieldDto.AiredStart,
 	[AnimeSortField.Status]: AnimeSortFieldDto.Status,
 	[AnimeSortField.None]: AnimeSortField.None,
+};
+
+/** Anime type transformation object in dto. */
+export const ANIME_TYPE_TO_DTO = {
+	[AnimeType.TV]: AnimeTypeDto.TV,
+	[AnimeType.OVA]: AnimeTypeDto.OVA,
+	[AnimeType.Movie]: AnimeTypeDto.Movie,
+	[AnimeType.Special]: AnimeTypeDto.Special,
+	[AnimeType.ONA]: AnimeTypeDto.ONA,
+	[AnimeType.Music]: AnimeTypeDto.Music,
+	[AnimeType.Unknown]: AnimeTypeDto.Unknown,
 };
 
 export namespace AnimeQueryParamsMapper {
